feat: Add common interface for findAll options and response
The `IFindAllOptions` interface is added to the `src/common/base/interface/common.interface.ts` file. It includes optional properties for `take` and `skip`, which can be used for pagination in the application.

The `IFindAllResponse` interface is also added to the same file. It includes properties for `data`, `total`, `take`, and `skip`, which are used to represent the response of a findAll operation.

These interfaces will be used to standardize the options and response format for findAll operations across the application.