feat: Add SuccessResponseInterceptor for transforming success responses
The `SuccessResponseInterceptor` class is added to the `src/common/response_service/interceptor/success_response.interceptor.ts` file. It implements the `NestInterceptor` interface from the `@nestjs/common` package and is responsible for transforming success responses in the application.

This interceptor modifies the response object by setting the status code and adding additional properties such as the timestamp and request path. It also transforms any `bigint` values in the response to strings.

This interceptor improves the consistency and customization of success responses in the application.