feat: Add IUpdateUserResponse interface for updating user information
This commit adds the `IUpdateUserResponse` interface to the `user.common.interfaces.ts` file in the `src/modules/user/application/interfaces` directory. The interface includes properties for `oldUser` and `newUser`, representing the previous and updated user objects respectively.

These changes enable the application to have a standardized interface for representing the response of updating user information.