feat: Add IUserController interface for updating user information
This commit adds the `IUserController` interface to the `user.controller.interface.ts` file in the `src/modules/user/application/interfaces` directory. The interface includes a method `updateUser` that takes an `UpdateUserDto` as a parameter and returns a promise response of type `IUpdateUserResponse`.

These changes enable the application to have a standardized interface for updating user information in the user controller.