feat: Add ICaslAbilityFactory interface for creating CaslAbility objects
This commit adds the `ICaslAbilityFactory` interface to the `casl-ability.interface.ts` file in the `src/modules/authorization/application/interface` directory. The interface defines a method `createForUser` that takes in a `User` object and returns an `AppAbility` object. This interface is responsible for creating CaslAbility objects based on the user's permissions and roles.

These changes enable the application to have a standardized way of creating CaslAbility objects, improving code organization and reusability.