feat: Add CaslAbilityFactory for creating CaslAbility objects
This commit adds the `CaslAbilityFactory` class to the `casl-ability.factory.ts` file in the `src/modules/authorization/infraestructure/casl/factory` directory. The `CaslAbilityFactory` implements the `ICaslAbilityFactory` interface and is responsible for creating `AppAbility` objects based on the user's permissions and roles.

These changes enable the application to have a standardized way of creating CaslAbility objects, improving code organization and reusability.