feat: Add AppAbility type for defining application abilities
This commit adds the `AppAbility` type to the `app-ability.type.ts` file in the `src/modules/authorization/infraestructure/casl/type` directory. The `AppAbility` type is a specific implementation of the `MongoAbility` type from the `@casl/ability` library. It defines the possible abilities that a user can have in the application, represented by a combination of `AppAction` and `AppSubjects`.

These changes enable the application to have a standardized way of defining and managing user abilities, improving code organization and maintainability.