feat: Add AppSubjects type for defining application subjects
This commit adds the `AppSubjects` type to the `app-subjects.type.ts` file in the `src/modules/authorization/infraestructure/casl/type` directory. The `AppSubjects` type is used to define the possible subjects that can be authorized in the application, including the `User` domain entity and the string literal `'all'`.

These changes enable the application to have a standardized way of defining and managing subjects for authorization, improving code organization and maintainability.