feat: Add permissions definition interface
This commit adds the `IPermissionsDefinition` interface to the `permissions-definition.interface.ts` file in the `src/modules/authorization/infraestructure/policy/type` directory. The `IPermissionsDefinition` interface is a partial record type that defines the permissions for each `AppRole` in the application. This change enables the application to have a standardized way of defining and managing permissions, improving code organization and maintainability.