feat: Add DefinePermissions type for defining permissions
This commit adds the `DefinePermissions` type to the `define-permissions.type.ts` file in the `src/modules/authorization/infraestructure/policy/type` directory. The `DefinePermissions` type is a function type that takes a `User` object and an `AbilityBuilder` object as parameters and defines the permissions for the user using the builder. This change enables the application to have a standardized way of defining permissions, improving code organization and maintainability.