feat: Update SuccessResponseInterceptor to handle optional statusCode
The SuccessResponseInterceptor in `success_response.interceptor.ts` has been updated to handle the `statusCode` property as optional in the `data` object. This change allows for more flexibility when setting the status code of the response. The interceptor will now only set the status code if it is provided in the `data` object.