Refinement of user registration and login functionalities
Enhancements have been rolled out for user registration and login services. Notably, the process now supports the registration of users already in the Cognito User Pool, adding this data to the local user storage if not already present there.

During login, if a user is not found in local storage but exists in Cognito and has valid credentials, the user is now automatically registered locally. Also, in alignment with this change, the ID token has been added to both the login and registration return payloads.

These changes improve the resilience of user management especially in scenarios involving data mismatches between local user data and the cloud-based Cognito User Pool.