Add getUserSub method to CognitoService for retrieving user sub attribute
Extended CognitoService in the AWS SDK with a new method, getUserSub, for directly retrieving the 'sub' attribute of user profiles given the email. Now, instead of calling multiple methods, we can fetch the 'sub' attribute using just one straightforward function. This streamlines our workflow and improves code readability. Also, to handle user not found errors, specific error handling was introduced in the getUserSub method.

Along with this, a minor correction in the registerUser condition block was made to prevent unintended executions during testing and development modes.