Removed isVerified and password fields from User model
Simplified the User model by removing the 'isVerified' and 'password' fields. These changes were made across all levels, from the DTO definition to the database schema. Regarding the 'password' field, it seems it was no longer necessary in the context of our current app. As for 'isVerified', we've changed our verification strategy, rendering this field redundant.

With these changes, the User model is now leaner and easier to manage. Please note that this modification has led to minor changes in the user service's response since the previously existing fields are no longer available, affecting the response body.

This commit presents a significant step towards a more streamlined application, improving maintainability and reducing potential points of failure. Consequently, it may impact related functionalities, so thorough testing is recommended to ensure the application's sound operation.