Added password change method to Cognito service
With increased concern for user security, we now allowed users to change their passwords while they are logged in. To support this feature, we've added a 'changePassword' method in the 'ICognitoAuthService' interface. This method expects an instance of 'ChangePasswordDto' and returns a promise that resolves to void. As this is an interface change, please ensure that implementations of this interface are updated accordingly.