Added change password functionality to Auth Controller
It was observed the Authentication Controller was handling most of the authentication-related tasks well, but it lacked the ability to change passwords. To address this, the code for the change password feature was added to the Auth Controller. This enhancement utilizes the 'ChangePasswordDto' data transfer object. The new function in the Auth Controller can now call the 'changePassword' function from the authentication service correctly with the necessary parameters.

This addition enhances the user experience by empowering users to change their password when necessary, thus providing a more comprehensive authentication system.