Add new endpoint for network status in System Status module
Introduced a new HTTP GET endpoint at 'system_status/soroban_network' for checking the status of the different networks in the System Status module. This makes it easier for the consumer services to fetch and monitor the status of the 'futureNetwork', 'testNetwork', and 'mainNetwork'. It uses the existing service 'SystemStatusService' for retrieving the network status. The endpoint is secured under 'Bearer' auth type to ensure expected accessibility.