Feat / Deploy Wasm File (#140)
## Summary

This PR introduces several updates related to the Stellar service and
invocation management. Key changes include the addition of new enums and
interfaces, modifications to existing service methods, and improvements
to the `InvocationService` for handling WASM files. It also includes new
endpoints for WASM file upload and preparation in the
`InvocationUserController`. Additionally, the extension size has
increased due to the required installation of Multer types, resulting in
updates to `package.json` and `package-lock.json`.