Fix / Deploy contracts SDK (#143)
## Summary

This PR introduces several improvements to the Stellar service and its
related modules. It focuses on refactoring the service to increase
modularity, enhance error handling, and add new features to improve the
handling of Stellar transactions, particularly in the context of
uploading WebAssembly (WASM) contracts. It also introduces a deployment
toggle option to the Invocation Service.