Fix/adapter server validation (#185)
This pull request includes several changes to the `docker-compose.yml`
and various files in the `src/common` directory to improve the codebase
and functionality. The most important changes include updates to volume
paths in `docker-compose.yml`, imports and test cases in
`stellar.service.spec.ts`, and method parameters in
`cognito.service.ts`.